requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==2.1.2
yfinance==0.2.50
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
